Added features to displaying data.
File AttributeDomain.js is responsible for creating domain for the attribute when the selected value type is "enumeration".

File ColumnHeaderMenu.js is responsible for the right click menu on header column and thus changing the column (attribute) properties such as if  attribute is active then there is the option to set it to inactive.

File Notification.js is used responsible for displaying quick pop-ups when something went wrong, right or informational ones. It is used in adding attribute to tell the user that he passed incorrect data (or didn't pass the data at all).

To the files EditData....js only classes were added/removed. Thanks to it - it is possible  to resize the window with buttons floating one below another (the file responsible for css is EditDataButtons.css).

File DropDownForAttributes.js as its name suggest is responsible for selecting value from drop down. It is used in adding attribute in many fields.

To the file DisplayData.js has been added such features as right click on column header, adding attribute or saving changes.