Edit attribute, save data, send data
In the file DisplayData.js the following functionality has been added: editing attribute, saving file on the computer, sending changes to the server.
In the file EditDataButtons.js only method name has been changed.
In the files DropDownForAttributes.js and AttributeDomain.js
has been added default values and also parent methods are executed. This way parent knows the value of the child before "handleChange" method. It is used in the edit attribute feature.