Make available to download ordinal misclassification matrix
Get request on path '/projects/{id}/misclassificationMatrix/download' allows to save choosen matrix object to text file. Serialization of misclassification matrix belongs to ruleLearn library and at the moment it is not implemented yet.

There are 4 types of ordinal misclassification matrices, which can be downloaded in current project:
- classification: matrix constructed at classification of data with given set of rules
- crossValidationSum: matrix of summarized statistics of all folds in cross-validation
- crossValidationMean: matrix of mean statistics of all folds in cross-validation
- crossValidationFold: matrix with classification result of given fold

There is an example of use in file 'GetMisclassificationMatrixDownload.js'.