Create GET and POST request on path '(...)/cones/descriptiveAttributes'
GET request returns json with array of available descriptive attributes and 'actual' field with attribute which is currently in use. If objects are described with default label, then field 'actual' is set to null.

POST request sets the attribute, which is currently used to describe objects. Name of the actual attribute is sent as a query parameter 'objectVisibleName'. If POST request doesn't contain this parameter, then default description is set to use. Received response has identical structure as in GET request.