Rewrite procedure of setting rules in project
Changes was forced by manipulating state of learning data in classification while rules in project are changing. It turns out that learning data flag in classification has  no connection with rules learning data, so functionality remains the same. Remember: isCurrentLearningData flag in project's classification tells if learning data of rules used in classification is the same as current data in 'DATA' tab.