Add new flag to the response on descriptive attributes of project
There is a new flag called "isEverywhere". If its value is 'true', then all particular resources have their descriptive attribute set on the same value. This flag is attached to the responses on GET/POST 'projects/{id}/metadata' requests. In order to achieve that, there is created a new class of response 'GlobalDescriptiveAttributesResponse'.